\name{NineteenEightyR-package}
\alias{NineteenEightyR-package}
\alias{NineteenEightyR}
\docType{package}
\title{R from another time and place.}
\description{Right now this just has some nice palettes and onload ascii art. Depends on colourlovers package to grab some palettes. Feel free to contribute!}
\author{
Sloth
}
\references{
Michael Mann, Lazerhawk, 288 GTO, Ray-Bans, Red Dawn, Commodore 64, Com Truise, Ice Pirates, Ocean Pacific, Stone washed denim, Atari 2600, that part where Goose dies, Dirk Daring, Walkman, Sloth from the Goonies, Turbo Kid, and many other things from both a distant past and the present that have an aesthetic that aliens discovering this planet in the future will totally understand.
}
\keyword{ package }
%\seealso{}
\examples{
'
           ^^                   @@@@@@@@@
      ^^       ^^            @@@@@@@@@@@@@@@
                           @@@@@@@@@@@@@@@@@@              ^^
                           @@@@@@@@@@@@@@@@@@@@
 ~~~~ ~~ ~~~~~ ~~~~~~~~ ~~ &&&&&&&&&&&&&&&&&&&& ~~~~~~~ ~~~~~~~~~~~ ~~~
 ~         ~~   ~  ~       ~~~~~~~~~~~~~~~~~~~~ ~       ~~     ~~ ~
   ~      ~~      ~~ ~~ ~~  ~~~~~~~~~~~~~ ~~~~  ~     ~~~    ~ ~~~  ~ ~~
   ~  ~~     ~         ~      ~~~~~~  ~~ ~~~       ~~ ~ ~~  ~~ ~
 ~  ~       ~ ~      ~           ~~ ~~~~~~  ~      ~~  ~             ~~
       ~             ~        ~      ~      ~~   ~             ~


'
}

