% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{sonny}
\alias{sonny}
\alias{malibu}
\alias{hotpink}
\alias{youngturqs}
\alias{sunset1}
\alias{sunset2}
\alias{sunset3}
\alias{miami1}
\alias{miami2}
\alias{seventies_aint_done_yet}
\alias{cobra}
\alias{electronic_night}
\title{Color palettes}
\usage{
sonny(n = 5)

malibu(n = 5)

hotpink(n = 5)

youngturqs(n = 12)

sunset1(n = 5)

sunset2(n = 5)

sunset3(n = 5)

miami1(n = 5)

miami2(n = 5)

seventies_aint_done_yet(n = 5)

cobra(n = 6)

electronic_night(n = 5)
}
\arguments{
\item{n}{The number of colors to return. The default is the length of the palette.}
}
\value{
A character vector of hex values for colors.  By default it will show the entire palette, otherwise the first 'n' or, it will repeat the palette to reach 'n' length.
}
\description{
Color palettes from another time and place.
}
\details{
This function returns the hex values of various colors.
}
\examples{
library(NineteenEightyR); library(ggplot2)
qplot(x=seq_along(sonny()), y='') + geom_tile(fill=sonny())
qplot(x=seq_along(malibu()), y='') + geom_tile(fill=malibu())
qplot(x=seq_along(seventies_aint_done_yet()), y='') + geom_tile(fill=seventies_aint_done_yet())
qplot(x=seq_along(youngturqs()), y='') + geom_tile(fill=youngturqs())
qplot(x=seq_along(cobra()), y='') + geom_tile(fill=cobra())
}
\seealso{
\code{\link[grDevices]{colors}}
}
