% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{sonny}
\alias{hotpink}
\alias{malibu}
\alias{sonny}
\alias{youngturqs}
\title{Color palettes}
\usage{
sonny(n = 8)

malibu(n = 8)

hotpink(n = 5)

youngturqs(n = 12)
}
\arguments{
\item{n}{The number of colors to return. The default is the length of the palette.}
}
\value{
A character vector of hex values for colors.
}
\description{
Color palettes from another time and place.
}
\details{
This function returns the hex values of various colors.
}
\examples{
library(NineteenEightyR)
ggplot2::qplot(x=seq_along(sonny()), fill=sonny(), geom='tile', y='')
}
\seealso{
\code{\link[grDevices]{colors}}
}

